admin done

+++ Controllers/ProjectsCRUD.cs
@@ -1,9 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
-using TaskMgmt.Auth;
 using TaskMgmt.Context;
 using TaskMgmt.Models;
 
@@ -18,12 +13,19 @@ namespace TaskMgmt.Controllers
             {
                 Console.WriteLine("Enter the name of the project:");
                 string name = Console.ReadLine();
+                Console.WriteLine("Enter the description of the project:");
+                string description = Console.ReadLine();
                 Project project = new Project
                 {
-                    Name = name
+                    Name = name,
+                    Description = description
                 };
                 db.Projects.Add(project);
                 db.SaveChanges();
+                const string message = "Project created successfully!";
+                Console.ForegroundColor = ConsoleColor.Green;
+                Console.WriteLine(message);
+                Console.ResetColor();
             }
         }
 
@@ -34,11 +36,27 @@ namespace TaskMgmt.Controllers
             {
                 Console.WriteLine("Enter the id of the project you want to update:");
                 int id = Convert.ToInt32(Console.ReadLine());
-                Console.WriteLine("Enter the new name of the project:");
-                string name = Console.ReadLine();
-                Project project = db.Projects.Find(id);
-                project.Name = name;
-                db.SaveChanges();
+                try
+                {
+                    var project = db.Projects.Find(id);
+                    Console.Write($"Old name: {project.Name}, Please enter the new name of the project:");
+                    string name = Console.ReadLine();
+                    Console.Write($"Old description: {project.Description}, Please enter the new description of the project:");
+                    string description = Console.ReadLine();
+                    project.Name = name;
+                    project.Description = description;
+                    db.SaveChanges();
+
+                    string message = "Project updated successfully!";
+                    Console.ForegroundColor = ConsoleColor.Green;
+                    Console.WriteLine(message);
+                    Console.ResetColor();
+
+                }
+                catch (Exception)
+                {
+                    throw;
+                }
             }
         }
 
@@ -50,8 +68,27 @@ namespace TaskMgmt.Controllers
                 Console.WriteLine("Enter the id of the project you want to delete:");
                 int id = Convert.ToInt32(Console.ReadLine());
                 Project project = db.Projects.Find(id);
-                db.Projects.Remove(project);
-                db.SaveChanges();
+                // are you sure you want to delete the project?
+                Console.WriteLine("Are you sure you want to delete the project? (y/n)");
+                string choice = Console.ReadLine();
+                if (choice == "y || Y || yes || Yes")
+                {
+                    // delete the project
+                    db.Projects.Remove(project);
+                    db.SaveChanges();
+                    // you have successfully deleted the project
+                    const string message = "Project deleted successfully!";
+                    Console.ForegroundColor = ConsoleColor.Green;
+                    Console.WriteLine(message);
+                    Console.ResetColor();
+                    Console.ReadKey();
+                }
+                else
+                {
+                    // do not delete the project
+                    Console.WriteLine("Project not deleted!");
+                    Console.ReadKey();
+                }
             }
         }
 
@@ -60,17 +97,43 @@ namespace TaskMgmt.Controllers
         {
             using (var db = new TaskContext())
             {
+                // get all the projects from the database to choose under which project the task should be created
+                List<Project> projects = db.Projects.ToList();
+                foreach (Project project in projects)
+                {
+                    Console.ForegroundColor = ConsoleColor.Yellow;
+                    Console.WriteLine($"Id: {project.Id}");
+                    Console.ForegroundColor = ConsoleColor.Green;
+                    Console.WriteLine($"Name: {project.Name}");
+                    Console.WriteLine($"Description: {project.Description}");
+                    Console.ResetColor();
+                }
+                Console.WriteLine("Enter the project id under which you want to create the task:");
+                int projectId = Convert.ToInt32(Console.ReadLine());
                 Console.WriteLine("Enter the title of the task:");
                 string title = Console.ReadLine();
                 Console.WriteLine("Enter the description of the task:");
                 string description = Console.ReadLine();
-                Console.WriteLine("Enter the project id of the task:");
-                int projectId = Convert.ToInt32(Console.ReadLine());
 
                 User assignedUser = null;
                 while (assignedUser == null)
                 {
-                    Console.WriteLine("Enter the assigned user id of the task:");
+                    // get all the users from the database to choose which user to assign the task with their status if they have a task assigned or not
+                    List<User> users = db.Users.Include(u => u.Task).ToList();
+                    foreach (User user in users)
+                    {
+                        Console.ForegroundColor = ConsoleColor.Yellow;
+                        Console.WriteLine($"Id: {user.Id}");
+                        Console.ForegroundColor = ConsoleColor.Green;
+                        Console.WriteLine($"Username: {user.Username}");
+                        Console.WriteLine($"Role: {user.Role}");
+                        if (user.Task != null)
+                        {
+                            Console.WriteLine($"Task: {user.Task.Title}");
+                        }
+                        Console.ResetColor();
+                    }
+                    Console.WriteLine("Enter the user id to assign the task:");
                     int assignedUserId = Convert.ToInt32(Console.ReadLine());
 
                     // Check if the user is already assigned a task
@@ -81,6 +144,19 @@ namespace TaskMgmt.Controllers
                         Console.WriteLine("User already has a task assigned! Choose another user.");
                         Console.ResetColor();
                         assignedUser = null; // Reset assignedUser to null to loop again
+                    }
+                    // if the user has a role of admin, show a message that admin cannot be assigned a task
+                    else if (assignedUser != null && assignedUser.Role == Role.Admin)
+                    {
+                        Console.ForegroundColor = ConsoleColor.Red;
+                        Console.WriteLine("Admin cannot be assigned a task! Choose another user.");
+                        Console.ResetColor();
+                        assignedUser = null; // Reset assignedUser to null to loop again
+                    }
+                    else
+                    {
+                        // if the user is not assigned a task, assign the task to the user
+                        assignedUser = db.Users.Find(assignedUserId);
                     }
                 }
 
@@ -103,7 +179,31 @@ namespace TaskMgmt.Controllers
         }
 
 
-        // View all tasks by status
+        // View all tasks
+        public void ViewAllTasks()
+        {
+            using (var db = new TaskContext())
+            {
+                List<Tasks> tasks = db.Tasks.ToList();
+                foreach (Tasks task in tasks)
+                {
+                    Console.ForegroundColor = ConsoleColor.Yellow;
+                    Console.WriteLine($"Id: {task.Id}");
+                    Console.ForegroundColor = ConsoleColor.Green;
+                    Console.WriteLine($"Title: {task.Title}");
+                    Console.WriteLine($"Description: {task.Description}");
+                    Console.WriteLine($"Status: {task.Status}");
+                    Console.ResetColor();
+                }
+                // call the method to view all tasks by status
+                Console.WriteLine("Do you want to view all tasks by status? (y/n)");
+                string choice = Console.ReadLine();
+                if (choice == "y")
+                {
+                    ViewAllTasksByStatus();
+                }
+            }
+        }
         // ViewAllTasksByStatus
         public void ViewAllTasksByStatus()
         {
@@ -127,7 +227,6 @@ namespace TaskMgmt.Controllers
                     Console.WriteLine($"Description: {task.Description}");
                     Console.WriteLine($"Status: {task.Status}");
                     Console.ResetColor();
-                    // Console.WriteLine($"Assigned to: {task.AssignedUser.Username}");
                 }
 
             }
@@ -140,23 +239,21 @@ namespace TaskMgmt.Controllers
             {
                 Console.WriteLine("Enter the id of the task you want to update:");
                 int id = Convert.ToInt32(Console.ReadLine());
-                Console.WriteLine("Enter the new title of the task:");
-                string title = Console.ReadLine();
-                Console.WriteLine("Enter the new description of the task:");
-                string description = Console.ReadLine();
-                Console.WriteLine("Enter the new status of the task:");
-                string status = Console.ReadLine();
-                Console.WriteLine("Enter the new project id of the task:");
-                int projectId = Convert.ToInt32(Console.ReadLine());
-                Console.WriteLine("Enter the new assigned user id of the task:");
-                int assignedUserId = Convert.ToInt32(Console.ReadLine());
-                Tasks task = db.Tasks.Find(id);
-                task.Title = title;
-                task.Description = description;
-                task.Status = (Status)Enum.Parse(typeof(Status), status);
-                task.ProjectId = projectId;
-                task.AssignedUserId = assignedUserId;
-                db.SaveChanges();
+                try
+                {
+                    var task = db.Tasks.Find(id);
+                    Console.Write($"Old title: {task.Title}, Please enter the new title of the task:");
+                    string title = Console.ReadLine();
+                    Console.Write($"Old description: {task.Description}, Please enter the new description of the task:");
+                    string description = Console.ReadLine();
+                    task.Title = title;
+                    task.Description = description;
+                    db.SaveChanges();
+                }
+                catch (Exception)
+                {
+                    throw;
+                }
             }
         }
 
@@ -165,11 +262,29 @@ namespace TaskMgmt.Controllers
         {
             using (var db = new TaskContext())
             {
+                // View all task and allow user to choose which task to delete
+                List<Tasks> tasks = db.Tasks.ToList();
+                foreach (Tasks task in tasks)
+                {
+                    Console.ForegroundColor = ConsoleColor.Yellow;
+                    Console.WriteLine($"Id: {task.Id}");
+                    Console.ForegroundColor = ConsoleColor.Green;
+                    Console.WriteLine($"Title: {task.Title}");
+                    Console.WriteLine($"Description: {task.Description}");
+                    Console.WriteLine($"Status: {task.Status}");
+                    Console.ResetColor();
+                }
                 Console.WriteLine("Enter the id of the task you want to delete:");
                 int id = Convert.ToInt32(Console.ReadLine());
-                Tasks task = db.Tasks.Find(id);
-                db.Tasks.Remove(task);
+                Tasks taskToDelete = db.Tasks.Find(id);
+                db.Tasks.Remove(taskToDelete);
                 db.SaveChanges();
+                // you have successfully deleted the task
+                const string message = "Task deleted successfully!";
+                Console.ForegroundColor = ConsoleColor.Green;
+                Console.WriteLine(message);
+                Console.ResetColor();
+                Console.ReadKey();
             }
         }
 
@@ -178,11 +293,27 @@ namespace TaskMgmt.Controllers
         {
             using (var db = new TaskContext())
             {
+                // View all users and allow user to choose which user to delete
+                List<User> users = db.Users.ToList();
+                foreach (User user in users)
+                {
+                    Console.ForegroundColor = ConsoleColor.Yellow;
+                    Console.WriteLine($"Id: {user.Id}");
+                    Console.ForegroundColor = ConsoleColor.Green;
+                    Console.WriteLine($"Username: {user.Username}");
+                    Console.WriteLine($"Role: {user.Role}");
+                    Console.ResetColor();
+                }
                 Console.WriteLine("Enter the id of the user you want to delete:");
                 int id = Convert.ToInt32(Console.ReadLine());
-                User user = db.Users.Find(id);
-                db.Users.Remove(user);
+                User userToDelete = db.Users.Find(id);
+                db.Users.Remove(userToDelete);
                 db.SaveChanges();
+                const string message = "User deleted successfully!";
+                Console.ForegroundColor = ConsoleColor.Green;
+                Console.WriteLine(message);
+                Console.ResetColor();
+                Console.ReadKey();
             }
         }
